test: 테스트용 모킹 유틸리티 함수 분리
